Fix L2 scratch memory allocation issue
Since `HBitVector` requests each L2 chunk with proper alignment
according to space alignment value, if the size of the vector is not
multiple of alignment the space taken by each team's bitvector would be
bigger than allocated resulting a memory allocation or illegal access
error.

This commit fixes the total allocation amount such that the space padded
for alignment is considered.